--- a/src/multisearch.rs
+++ b/src/multisearch.rs
@@ -77,25 +77,27 @@ pub fn multisearch(
                 let target_size = against.minhash.size() as f64;
 
                 let containment_query_in_target = overlap / query_size;
-                let containment_target_in_query = overlap / target_size;
-                let max_containment = containment_query_in_target.max(containment_target_in_query);
-                let jaccard = overlap / (target_size + query_size - overlap);
-
-                // estimate ANI values
-                let mut query_ani = None;
-                let mut match_ani = None;
-                let mut average_containment_ani = None;
-                let mut max_containment_ani = None;
-                if estimate_ani {
-                    let qani = ani_from_containment(containment_query_in_target, ksize) * 100.0;
-                    let mani = ani_from_containment(containment_target_in_query, ksize) * 100.0;
-                    query_ani = Some(qani);
-                    match_ani = Some(mani);
-                    average_containment_ani = Some((qani + mani) / 2.);
-                    max_containment_ani = Some(f64::max(qani, mani));
-                }
 
                 if containment_query_in_target > threshold {
+                    let containment_target_in_query = overlap / target_size;
+                    let max_containment =
+                        containment_query_in_target.max(containment_target_in_query);
+                    let jaccard = overlap / (target_size + query_size - overlap);
+                    let mut query_ani = None;
+                    let mut match_ani = None;
+                    let mut average_containment_ani = None;
+                    let mut max_containment_ani = None;
+
+                    // estimate ANI values
+                    if estimate_ani {
+                        let qani = ani_from_containment(containment_query_in_target, ksize) * 100.0;
+                        let mani = ani_from_containment(containment_target_in_query, ksize) * 100.0;
+                        query_ani = Some(qani);
+                        match_ani = Some(mani);
+                        average_containment_ani = Some((qani + mani) / 2.);
+                        max_containment_ani = Some(f64::max(qani, mani));
+                    }
+
                     results.push(MultiSearchResult {
                         query_name: query.name.clone(),
                         query_md5: query.md5sum.clone(),