--- a/src/python/tests/test_multisearch.py
+++ b/src/python/tests/test_multisearch.py
@@ -66,6 +66,10 @@ def test_simple(runtmp, zip_query, zip_db):
             assert float(row['containment'] == 1.0)
             assert float(row['jaccard'] == 1.0)
             assert float(row['max_containment'] == 1.0)
+            assert float(row['query_ani'] == 100.0)
+            assert float(row['match_ani'] == 100.0)
+            assert float(row['average_containment_ani'] == 100.0)
+            assert float(row['max_containment_ani'] == 100.0)
 
         else:
             # confirm hand-checked numbers
@@ -75,23 +79,40 @@ def test_simple(runtmp, zip_query, zip_db):
             jaccard = float(row['jaccard'])
             maxcont = float(row['max_containment'])
             intersect_hashes = int(row['intersect_hashes'])
+            q1_ani = float(row['query_ani'])
+            q2_ani = float(row['match_ani'])
+            avg_ani = float(row['average_containment_ani'])
+            max_ani = float(row['max_containment_ani'])
+
 
             jaccard = round(jaccard, 4)
             cont = round(cont, 4)
             maxcont = round(maxcont, 4)
-            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}")
+            q1_ani = round(q1_ani, 2)
+            q2_ani = round(q2_ani, 2)
+            avg_ani = round(avg_ani, 2)
+            max_ani = round(max_ani, 2)
+            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", f"{q1_ani:.04}", f"{q2_ani:.04}", f"{avg_ani:.04}", f"{max_ani:.04}")
 
             if q == 'NC_011665.1' and m == 'NC_009661.1':
                 assert jaccard == 0.3207
                 assert cont == 0.4828
                 assert maxcont == 0.4885
                 assert intersect_hashes == 2529
+                assert q1_ani == 97.68
+                assert q2_ani == 97.72
+                assert avg_ani == 97.7
+                assert max_ani == 97.72
 
             if q == 'NC_009661.1' and m == 'NC_011665.1':
                 assert jaccard == 0.3207
                 assert cont == 0.4885
                 assert maxcont == 0.4885
                 assert intersect_hashes == 2529
+                assert q1_ani == 97.72
+                assert q2_ani == 97.68
+                assert avg_ani == 97.7
+                assert max_ani == 97.72
 
 
 @pytest.mark.parametrize("zip_query", [False, True])
@@ -512,6 +533,10 @@ def test_simple_prot(runtmp):
             assert float(row['containment'] == 1.0)
             assert float(row['jaccard'] == 1.0)
             assert float(row['max_containment'] == 1.0)
+            assert float(row['query_ani'] == 100.0)
+            assert float(row['match_ani'] == 100.0)
+            assert float(row['average_containment_ani'] == 100.0)
+            assert float(row['max_containment_ani'] == 100.0)
 
         else:
             # confirm hand-checked numbers
@@ -521,23 +546,39 @@ def test_simple_prot(runtmp):
             jaccard = float(row['jaccard'])
             maxcont = float(row['max_containment'])
             intersect_hashes = int(row['intersect_hashes'])
+            q1_ani = float(row['query_ani'])
+            q2_ani = float(row['match_ani'])
+            avg_ani = float(row['average_containment_ani'])
+            max_ani = float(row['max_containment_ani'])
 
             jaccard = round(jaccard, 4)
             cont = round(cont, 4)
             maxcont = round(maxcont, 4)
-            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes)
+            q1_ani = round(q1_ani, 2)
+            q2_ani = round(q2_ani, 2)
+            avg_ani = round(avg_ani, 2)
+            max_ani = round(max_ani, 2)
+            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes, f"{q1_ani:.04}", f"{q2_ani:.04}", f"{avg_ani:.04}", f"{max_ani:.04}")
 
             if q == 'GCA_001593925' and m == 'GCA_001593935':
                 assert jaccard == 0.0434
                 assert cont == 0.1003
                 assert maxcont == 0.1003
                 assert intersect_hashes == 342
+                assert q1_ani == 88.6
+                assert q2_ani == 87.02
+                assert avg_ani == 87.81
+                assert max_ani == 88.6
 
             if q == 'GCA_001593935' and m == 'GCA_001593925':
                 assert jaccard == 0.0434
                 assert cont == 0.0712
                 assert maxcont == 0.1003
                 assert intersect_hashes == 342
+                assert q1_ani == 87.02
+                assert q2_ani == 88.6
+                assert avg_ani == 87.81
+                assert max_ani == 88.6
 
 
 def test_simple_dayhoff(runtmp):
@@ -564,6 +605,10 @@ def test_simple_dayhoff(runtmp):
             assert float(row['containment'] == 1.0)
             assert float(row['jaccard'] == 1.0)
             assert float(row['max_containment'] == 1.0)
+            assert float(row['query_ani'] == 100.0)
+            assert float(row['match_ani'] == 100.0)
+            assert float(row['average_containment_ani'] == 100.0)
+            assert float(row['max_containment_ani'] == 100.0)
 
         else:
             # confirm hand-checked numbers
@@ -573,23 +618,39 @@ def test_simple_dayhoff(runtmp):
             jaccard = float(row['jaccard'])
             maxcont = float(row['max_containment'])
             intersect_hashes = int(row['intersect_hashes'])
+            q1_ani = float(row['query_ani'])
+            q2_ani = float(row['match_ani'])
+            avg_ani = float(row['average_containment_ani'])
+            max_ani = float(row['max_containment_ani'])
 
             jaccard = round(jaccard, 4)
             cont = round(cont, 4)
             maxcont = round(maxcont, 4)
-            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes)
+            q1_ani = round(q1_ani, 2)
+            q2_ani = round(q2_ani, 2)
+            avg_ani = round(avg_ani, 2)
+            max_ani = round(max_ani, 2)
+            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes, f"{q1_ani:.04}", f"{q2_ani:.04}", f"{avg_ani:.04}", f"{max_ani:.04}")
 
             if q == 'GCA_001593925' and m == 'GCA_001593935':
                 assert jaccard == 0.1326
                 assert cont == 0.2815
                 assert maxcont == 0.2815
                 assert intersect_hashes == 930
+                assert q1_ani == 93.55
+                assert q2_ani == 91.89
+                assert avg_ani == 92.72
+                assert max_ani == 93.55
 
             if q == 'GCA_001593935' and m == 'GCA_001593925':
                 assert jaccard == 0.1326
                 assert cont == 0.2004
                 assert maxcont == 0.2815
                 assert intersect_hashes == 930
+                assert q1_ani == 91.89
+                assert q2_ani == 93.55
+                assert avg_ani == 92.72
+                assert max_ani == 93.55
 
 
 def test_simple_hp(runtmp):
@@ -616,6 +677,10 @@ def test_simple_hp(runtmp):
             assert float(row['containment'] == 1.0)
             assert float(row['jaccard'] == 1.0)
             assert float(row['max_containment'] == 1.0)
+            assert float(row['query_ani'] == 100.0)
+            assert float(row['match_ani'] == 100.0)
+            assert float(row['average_containment_ani'] == 100.0)
+            assert float(row['max_containment_ani'] == 100.0)
 
         else:
             # confirm hand-checked numbers
@@ -625,20 +690,36 @@ def test_simple_hp(runtmp):
             jaccard = float(row['jaccard'])
             maxcont = float(row['max_containment'])
             intersect_hashes = int(row['intersect_hashes'])
+            q1_ani = float(row['query_ani'])
+            q2_ani = float(row['match_ani'])
+            avg_ani = float(row['average_containment_ani'])
+            max_ani = float(row['max_containment_ani'])
 
             jaccard = round(jaccard, 4)
             cont = round(cont, 4)
             maxcont = round(maxcont, 4)
-            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes)
+            q1_ani = round(q1_ani, 2)
+            q2_ani = round(q2_ani, 2)
+            avg_ani = round(avg_ani, 2)
+            max_ani = round(max_ani, 2)
+            print(q, m, f"{jaccard:.04}", f"{cont:.04}", f"{maxcont:.04}", intersect_hashes, f"{q1_ani:.04}", f"{q2_ani:.04}", f"{avg_ani:.04}", f"{max_ani:.04}")
 
             if q == 'GCA_001593925' and m == 'GCA_001593935':
                 assert jaccard == 0.4983
                 assert cont == 0.747
                 assert maxcont == 0.747
                 assert intersect_hashes == 1724
+                assert q1_ani == 98.48
+                assert q2_ani == 97.34
+                assert avg_ani == 97.91
+                assert max_ani == 98.48
 
             if q == 'GCA_001593935' and m == 'GCA_001593925':
                 assert jaccard == 0.4983
                 assert cont == 0.5994
                 assert maxcont == 0.747
                 assert intersect_hashes == 1724
+                assert q1_ani == 97.34
+                assert q2_ani == 98.48
+                assert avg_ani == 97.91
+                assert max_ani == 98.48