--- a/src/utils.rs
+++ b/src/utils.rs
@@ -8,7 +8,7 @@ use camino::Utf8Path as Path;
 use camino::Utf8PathBuf as PathBuf;
 use csv::Writer;
 use serde::ser::Serializer;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use std::cmp::{Ordering, PartialOrd};
 use std::collections::BinaryHeap;
 use std::fs::{create_dir_all, File};
@@ -716,7 +716,7 @@ pub struct BranchwaterGatherResult {
     pub intersect_bp: usize,
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Deserialize)]
 pub struct MultiSearchResult {
     pub query_name: String,
     pub query_md5: String,
@@ -726,6 +726,10 @@ pub struct MultiSearchResult {
     pub max_containment: f64,
     pub jaccard: f64,
     pub intersect_hashes: f64,
+    pub query_ani: f64,
+    pub match_ani: f64,
+    pub average_containment_ani: f64,
+    pub max_containment_ani: f64,
 }
 
 #[derive(Serialize)]