@mr-eyes 2024-02-24
Hi @bluegenes 

Do you have benchmark results by chance? As in #188