@bluegenes 2024-02-24
@mr-eyes I benchmarked ANI as part of `gather` over in sourmash (https://github.com/sourmash-bio/sourmash/pull/2943), but yeah, happy to do it here too for a better comparison.

Do you have the sigs you used in #188 handy?