@mr-eyes 2024-02-24
> @mr-eyes I benchmarked this streamlined ANI as part of `gather` over in sourmash ([sourmash-bio/sourmash#2943](https://github.com/sourmash-bio/sourmash/pull/2943)), but yeah, would be good to do it here too for a better comparison.
> 

I think https://github.com/sourmash-bio/sourmash/pull/2943 is benchmarking Rust vs Python version. However, I meant to use the #188 benchmark to compare Rust vs Rust in branchwater. When I added ANI pairwise, the time was doubled. I think @ctb was surprised by the slowdown https://github.com/sourmash-bio/sourmash_plugin_branchwater/pull/188#issuecomment-1908166368, so I thought that `--ani` could be added as a flag, but that is more coding, etc. 
What do you think? 

> Do you have the sigs you used in #188 handy / would you be willing to share?
Unfortunately, it was local on my Ubuntu, and I crashed it two days ago :/ but I am happy to share some large signatures with you on the farm if you'd like to.