@bluegenes 2024-02-24
> I think https://github.com/sourmash-bio/sourmash/pull/2943 is benchmarking Rust vs Python version like in https://github.com/sourmash-bio/sourmash_plugin_branchwater/issues/164#issuecomment-1899175593. However, I meant to use the https://github.com/sourmash-bio/sourmash_plugin_branchwater/pull/188 benchmark to compare Rust vs Rust in branchwater. 

I know. I was thinking about my initial benchmarking of ANI, in which i didn't look like ANI added much time. But then I remembered that that wasn't actually running anything, and after a fix by luiz, ANI did add extra time 😂 . 

I did streamlined the calculations, but they won't be that much faster (if at all) than your #188 implementation.

> When I added ANI pairwise, the time was doubled. I think @ctb was surprised by the slowdown https://github.com/sourmash-bio/sourmash_plugin_branchwater/pull/188#issuecomment-1908166368, so I thought that --ani could be added as a flag, but that is more coding, etc.

hah, just did this prior to seeing this point. It was annoying but is done now 🤷🏻‍♀️ 

> Unfortunately, it was local on my Ubuntu, and I crashed it two days ago :/ but I am happy to share some large signatures with you on the farm if you'd like to.

no worries, i have some. was just trying to be comparable