@bluegenes 2024-02-24
With the 12k ICTV viral genomes, scaled 200, ANI doesn't add much time. 

The whole point of streamlining the ANI for `sourmash` core was to make ANI faster .. so this is nice? But also I'm suspicious and will also try some larger sigs.

No ANI:
```
DONE. Processed 79752135 comparisons
...pairwise is done! results in 'vmr_MSL38_v1.dna-k21-sc200.pairwise.csv'
        Command being timed: "sourmash scripts pairwise vmr_MSL38_v1.dna-k21-sc200.zip -c 10 -k 21 -s 200 -o vmr_MSL38_v1.dna-k21-sc200.pairwise.csv"
        User time (seconds): 156.95
        System time (seconds): 4.74
        Percent of CPU this job got: 769%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:21.01
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 171876
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 219
        Minor (reclaiming a frame) page faults: 95860
        Voluntary context switches: 137178
        Involuntary context switches: 158648
        Swaps: 0
        File system inputs: 54216
        File system outputs: 41112
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

with ANI:
```
DONE. Processed 79752135 comparisons
...pairwise is done! results in 'vmr_MSL38_v1.dna-k21-sc200.pairwise-ani.csv'
        Command being timed: "sourmash scripts pairwise vmr_MSL38_v1.dna-k21-sc200.zip -c 10 -k 21 -s 200 --ani -o vmr_MSL38_v1.dna-k21-sc200.pairwise-ani.csv"
        User time (seconds): 159.37
        System time (seconds): 4.81
        Percent of CPU this job got: 804%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:20.40
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 169372
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 94462
        Voluntary context switches: 132760
        Involuntary context switches: 149866
        Swaps: 0
        File system inputs: 0
        File system outputs: 54816
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```


using v0.9.0 (no ANI):
```
DONE. Processed 79752135 comparisons
...pairwise is done! results in 'vmr_MSL38_v1.dna-k21-sc200.pairwise-main.csv'
        Command being timed: "sourmash scripts pairwise vmr_MSL38_v1.dna-k21-sc200.zip -c 10 -k 21 -s 200 -o vmr_MSL38_v1.dna-k21-sc200.pairwise-main.csv"
        User time (seconds): 158.68
        System time (seconds): 4.55
        Percent of CPU this job got: 791%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:20.62
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 169864
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 107970
        Voluntary context switches: 133159
        Involuntary context switches: 174822
        Swaps: 0
        File system inputs: 0
        File system outputs: 41112
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```